bpo-21622: Fix ctypes.util.find_library with musl
Musl libc does not use any ld.cache and `ldconfig -r` does not work with
musl. It is also common that musl based container runtimes excludes
objdump, gcc and ld, which means that find_library() does not work at
all.

So as a last resort, if none of the previously mentioned methods works,
scan LD_LIBRARY_PATH and some standard locations for elf files that
matches the specified name, in a similar way that ld does.

We also update the tests so they work as expected with musl libc, which
does not have a separate libm or libcrypt.

Signed-off-by: Natanael Copa <ncopa@alpinelinux.org>